Extract href generator for derived value query
De-duplicates the generation as the links are identical for both
DisplayField modules.
