+++ frontend/src/app/shared/components/fields/display/field-types/compound-progress-display-field.module.ts
@@ -30,14 +30,15 @@ import { ApiV3Service } from 'core-app/core/apiv3/api-v3.service';
 import { DisplayField } from 'core-app/shared/components/fields/display/display-field.module';
 import { InjectField } from 'core-app/shared/helpers/angular/inject-field.decorator';
 import { PathHelperService } from 'core-app/core/path-helper/path-helper.service';
-import { ProjectResource } from 'core-app/features/hal/resources/project-resource';
 import { uiStateLinkClass } from 'core-app/features/work-packages/components/wp-fast-table/builders/ui-state-link-builder';
-import * as URI from 'urijs';
+import { HierarchyQueryLinkHelperService } from 'core-app/shared/components/fields/display/field-types/hierarchy-query-link-helper.service';
 
 export class CompoundProgressDisplayField extends DisplayField {
   @InjectField() PathHelper:PathHelperService;
 
   @InjectField() apiV3Service:ApiV3Service;
+
+  @InjectField() hierarchyQueryLinkHelper:HierarchyQueryLinkHelperService;
 
   private derivedText = this.I18n.t('js.label_value_derived_from_children');
 
@@ -70,7 +71,8 @@ export class CompoundProgressDisplayField extends DisplayField {
     link.title = `${displayText} ${this.derivedText}`;
     link.classList.add('-derived-value', uiStateLinkClass);
 
-    this.addURLToViewWorkPackageChildren(link);
+    this.hierarchyQueryLinkHelper.addHref(link, this.resource);
+    // this.addURLToViewWorkPackageChildren(link);
 
     element.appendChild(link);
   }
@@ -83,31 +85,6 @@ export class CompoundProgressDisplayField extends DisplayField {
     span.ariaHidden = 'true';
 
     element.appendChild(span);
-  }
-
-  private addURLToViewWorkPackageChildren(link:HTMLAnchorElement):void {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-    if (this.resource && this.resource.id && this.resource.project) {
-      const wpID = this.resource.id.toString();
-      this
-        .apiV3Service
-        .projects
-        .id(this.resource.project as ProjectResource)
-        .get()
-        .subscribe((project:ProjectResource) => {
-          const props = {
-            c: ['id', 'subject', 'type', 'status', 'estimatedTime', 'remainingTime', 'percentageDone'],
-            hi: true,
-            is: true,
-            f: [{ n: 'parent', o: '=', v: [wpID] }],
-          };
-          const href = URI(this.PathHelper.projectWorkPackagesPath(project.identifier as string))
-            .query({ query_props: JSON.stringify(props) })
-            .toString();
-
-          link.href = href;
-        });
-    }
   }
 
   public get value():number {
