+++ frontend/src/app/shared/components/fields/display/field-types/estimated-time-display-field.module.ts
@@ -28,12 +28,11 @@
 
 import { DisplayField } from 'core-app/shared/components/fields/display/display-field.module';
 import { InjectField } from 'core-app/shared/helpers/angular/inject-field.decorator';
-import * as URI from 'urijs';
 import { TimezoneService } from 'core-app/core/datetime/timezone.service';
-import { ProjectResource } from 'core-app/features/hal/resources/project-resource';
 import { ApiV3Service } from 'core-app/core/apiv3/api-v3.service';
 import { PathHelperService } from 'core-app/core/path-helper/path-helper.service';
 import { uiStateLinkClass } from 'core-app/features/work-packages/components/wp-fast-table/builders/ui-state-link-builder';
+import { HierarchyQueryLinkHelperService } from 'core-app/shared/components/fields/display/field-types/hierarchy-query-link-helper.service';
 
 export class EstimatedTimeDisplayField extends DisplayField {
   @InjectField() timezoneService:TimezoneService;
@@ -41,6 +40,8 @@ export class EstimatedTimeDisplayField extends DisplayField {
   @InjectField() PathHelper:PathHelperService;
 
   @InjectField() apiV3Service:ApiV3Service;
+
+  @InjectField() hierarchyQueryLinkHelper:HierarchyQueryLinkHelperService;
 
   private derivedText = this.I18n.t('js.label_value_derived_from_children');
 
@@ -115,7 +116,7 @@ export class EstimatedTimeDisplayField extends DisplayField {
     link.title = `${this.derivedValueString} ${this.derivedText}`;
     link.classList.add('-derived-value', uiStateLinkClass);
 
-    this.addURLToViewWorkPackageChildren(link);
+    this.hierarchyQueryLinkHelper.addHref(link, this.resource);
 
     element.appendChild(link);
   }
@@ -130,30 +131,5 @@ export class EstimatedTimeDisplayField extends DisplayField {
     const derived = this.derivedValue;
 
     return !value && !derived;
-  }
-
-  private addURLToViewWorkPackageChildren(link:HTMLAnchorElement):void {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-    if (this.resource && this.resource.id && this.resource.project) {
-      const wpID = this.resource.id.toString();
-      this
-        .apiV3Service
-        .projects
-        .id(this.resource.project as ProjectResource)
-        .get()
-        .subscribe((project:ProjectResource) => {
-          const props = {
-            c: ['id', 'subject', 'type', 'status', 'estimatedTime', 'remainingTime'],
-            hi: true,
-            is: true,
-            f: [{ n: 'parent', o: '=', v: [wpID] }],
-          };
-          const href = URI(this.PathHelper.projectWorkPackagesPath(project.identifier as string))
-            .query({ query_props: JSON.stringify(props) })
-            .toString();
-
-          link.href = href;
-        });
-    }
   }
 }
